Try to fetch tomorrow's prices, hopefully fixing the midnight bug (stuck at 23-0)
I'm guessing that just fetching until the end of today played unwell with the timeline provider and the .atEnd policy we gave it. I.e. in the span 23:00-24:00, where WidgetKit probably gives us the opportunity to update the timeline, we would just return the same (expired) entries again. Probably making WidgetKit think that we're not "behaving" and not giving us any more chances to update.