Add reading the security token in the extension targets, fixes #6
I had simply overlooked/forgot this scenario and that I hadn't added the security token to the extension targets.
Since the app and extension (the same principle applies to the watch and the iphone/ipad apps) share the data in a app group, in many cases (e.g. if opening the app at some point) it's the host app that makes the network request to refresh the data, and both app and extensions are good for the rest of the day (sometimes also the day after if having fetched tomorrow's prices). But when leaving the app for a while, letting the data "run out", it's the extension that's woken up by the OS to make the request. And without a token, that failed.