Remove Alamofire in favour of a custom SessionManager, supporting forground to background network task transitioning
The idea is that since foreground networking (data tasks) are much more performant, we want to prefer those. But if/when they take too long to finish, they will be cancelled by the system and we need to retry the request as a download task. The WWDC presentation "Efficiency awaits: Background tasks in SwiftUI" stated that the "URLSession will deduplicate any in-process requests under the hood." which sounds like a really nice feature. See:
- https://developer.apple.com/videos/play/wwdc2022/10142/?time=618
- https://developer.apple.com/forums/thread/709437
I haven't verified that this actually works yet or that I haven't misunderstood anything of the above.